Renamed drag and drop component folder
