+++ components/index.tsx
@@ -1,3 +1,3 @@
-export {default as DragAndDropSection} from './drag-and-drop-section/DragAndDropSection'
+export {default as DragAndDropSection} from './DragAndDropSelection/DragAndDropSection'
 export {default as Button} from './common/Button'
 export {default as Input} from './common/Input'
